# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/issues": {
            "get": {
                "description": "Get a list of all issues",
                "produces": [
                    "application/json"
                ],
                "summary": "List all issues",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Issue"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new issue",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new issue",
                "parameters": [
                    {
                        "description": "Issue object to be created",
                        "name": "issue",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Issue"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Issue"
                        }
                    }
                }
            }
        },
        "/issues/{id}": {
            "get": {
                "description": "Get an issue by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get an issue by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Issue ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Issue"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing issue",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update an existing issue",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Issue ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated issue object",
                        "name": "issue",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Issue"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Issue"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an issue by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Delete an issue by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Issue ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "patch": {
                "description": "Update specific fields of an existing issue",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update specific fields of an existing issue",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Issue ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to be updated",
                        "name": "updateRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Issue"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Issue"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Issue": {
            "description": "Issue Structure",
            "type": "object",
            "properties": {
                "assigned_to": {
                    "description": "name of assignee of the issue",
                    "type": "string"
                },
                "category": {
                    "description": "issue category",
                    "type": "string"
                },
                "created_by": {
                    "description": "name of user who creates issue",
                    "type": "string"
                },
                "created_on": {
                    "description": "issue created date",
                    "type": "string"
                },
                "eta": {
                    "description": "time whne issue will be fixed",
                    "type": "string"
                },
                "issue_description": {
                    "description": "issue descriptiom",
                    "type": "string"
                },
                "issue_id": {
                    "type": "integer"
                },
                "priority": {
                    "description": "issue prioroty",
                    "type": "string"
                },
                "state": {
                    "description": "issue state",
                    "type": "string"
                },
                "tags": {
                    "description": "tags to filter issue",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updated_on": {
                    "description": "issue updated time",
                    "type": "string"
                }
            }
        }
    }
}